backend/requirements.txt
fastapi
uvicorn
pydantic
google-genai
python-dotenv
requests
python-multipart
Pillow
